+++ src/sorter_modules/parallel_QA.py
@@ -23,16 +23,6 @@ FORMATTING_FUNCS = (
 )
 
 
-def printl(text=""):
-    """
-    Function for printing text to console and to log file.
-    """
-    print(text)
-    logfile.write(str(text).strip('\r') + '\n')
-    logfile.flush()
-# end def printl
-
-
 from src.sorter_modules.common import *